WIP: 0.7 Requires.jl code
