--- recession_dates.tsv
@@ -1,34 +1,0 @@
-1854-12-01	1855-01-01
-1857-07-01	1859-01-01
-1860-11-01	1861-07-01
-1865-05-01	1868-01-01
-1869-07-01	1871-01-01
-1873-11-01	1879-04-01
-1882-04-01	1885-06-01
-1887-04-01	1888-05-01
-1890-08-01	1891-06-01
-1893-02-01	1894-07-01
-1896-01-01	1897-07-01
-1899-07-01	1901-01-01
-1902-10-01	1904-09-01
-1907-06-01	1908-07-01
-1910-02-01	1912-02-01
-1913-02-01	1915-01-01
-1918-09-01	1919-04-01
-1920-02-01	1921-08-01
-1923-06-01	1924-08-01
-1926-11-01	1927-12-01
-1929-09-01	1933-04-01
-1937-06-01	1938-07-01
-1945-03-01	1945-11-01
-1948-12-01	1949-11-01
-1953-08-01	1954-06-01
-1957-09-01	1958-05-01
-1960-05-01	1961-03-01
-1970-01-01	1970-12-01
-1973-12-01	1975-04-01
-1980-02-01	1980-08-01
-1981-08-01	1982-12-01
-1990-08-01	1991-04-01
-2001-04-01	2001-12-01
-2008-01-01	2009-07-01
