+++ src/display.jl
@@ -122,8 +122,6 @@ Base.copy(sp::SyncPlot) = fork(sp)  # defined by each SyncPlot{TD}
 
 using Compat.UUIDs
 
-@require Juno include("displays/juno.jl")
-@require WebIO include("displays/webio.jl")
 include("displays/electron.jl")
 include("displays/ijulia.jl")
